refactor(nms): Cleanup state directory
