+++ nms/app/host.tsx
@@ -11,8 +11,8 @@
  * limitations under the License.
  */
 
-import './common/axiosConfig';
-import './common/polyfill';
+import './util/axiosConfig';
+import './util/polyfill';
 
 import Index from './components/host/Index';
 import React from 'react';
