+++ nms/app/views/alarms/AlarmsDashboard.tsx
@@ -22,11 +22,11 @@ import React from 'react';
 import Text from '../../theme/design-system/Text';
 import TopBar from '../../components/TopBar';
 import nullthrows from '../../../shared/util/nullthrows';
-import {MagmaAlarmsApiUtil} from '../../state/AlarmsApiUtil';
+import {MagmaAlarmsApiUtil} from '../../util/AlarmsApiUtil';
 import {Theme} from '@material-ui/core/styles';
 import {colors, typography} from '../../theme/default';
 import {makeStyles} from '@material-ui/styles';
-import {triggerAlertSync} from '../../state/SyncAlerts';
+import {triggerAlertSync} from '../../util/SyncAlerts';
 import {useContext} from 'react';
 import {useEnqueueSnackbar} from '../../hooks/useSnackbar';
 import {useParams} from 'react-router-dom';
